Address PR review comments
- Fix stat() to return 0o755 for directories (was always 0o644)
- Reject unknown compression types instead of silent acceptance
- Use decompress_into to avoid intermediate heap allocation
- Add envelope length validation in read_block()
- Increase verify_block_crc buffer with saturating arithmetic
- Change DirEntry.name_len from usize to u32 for ABI stability
- Fix mkdir doc comment to match implementation
- Update developer-guide no_std constraints (shared uses alloc)

+++ shared/src/storage.rs
@@ -746,8 +746,8 @@ pub struct PosixStat {
 pub struct DirEntry {
     /// Entry name (up to 64 bytes, null-padded).
     pub name: [u8; MAX_OBJECT_NAME_LEN],
-    /// Name length in bytes.
-    pub name_len: usize,
+    /// Name length in bytes (fixed-width for cross-boundary ABI stability).
+    pub name_len: u32,
     /// Object identifier.
     pub object_id: ObjectId,
     /// Content type.
